% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargeCalculationGlobal.R
\name{chargeCalculationGlobal}
\alias{chargeCalculationGlobal}
\title{Protein Charge Calculation, Globally}
\usage{
chargeCalculationGlobal(
  sequence,
  pKaSet = "IPC_protein",
  pH = 7.2,
  plotResults = FALSE,
  includeTermini = TRUE,
  sumTermini = TRUE,
  proteinName = NA,
  printCitation = FALSE,
  ...
)
}
\arguments{
\item{sequence}{amino acid sequence as a character string or vector of
individual residues. alternativly a character string of the path to a
.fasta / .fa file}

\item{pKaSet}{A character string or data frame. "IPC_protein" by default.
Character string to load specific, preloaded pKa sets.
 c("AA", "EMBOSS", "DTASelect", "Solomons", "Sillero", "Rodwell",
  "Lehninger", "Toseland", "Thurlkill", "Nozaki", "Dawson",
  "Bjellqvist", "ProMoST", "IPC_protein", "IPC_peptide")
 Alternativly, the user may supply a custom pKa dataset.
 The format must be a data frame where:
 Column 1 must be a character vector of residues AND
 Column 2 must be a numeric vector of pKa values.}

\item{pH}{numeic value, 7.2 by default.
The envioronmental pH used to calculate residue charge.}

\item{plotResults}{logical value, FALSE by default.
This determines what is returned. If \code{plotResults = FALSE}, a
dataframe is returned with the position, residue, and charge (-1 to +1). If
\code{plotResults = TRUE}, a graphical output is returned (ggplot) showing
the charge distribution.}

\item{includeTermini, sumTermini}{Logical values, both TRUE by default. This
determines how the calculation handles the N- and C- terminus.
includeTermini determines if the calculation will use the charge of the
amine and carboxyl groups at the ends of the peptide (When TRUE). These
charges are ignored when \code{includeTermini = FALSE}. sumTermini
determines if the charge of the first (likely Met, therefore uncharged),
and final residue (varies) will be added to the termini charges, or if the
N and C terminus will be returned as seperate residues.
When \code{sumTermini = TRUE}, charges are summed. When
\code{sumTermini = FALSE}, the N and C terminus are added as a unique
residue in the DF. This will impact averages by increasing the sequence
length by 2. sumTermini is ignored if \code{includeTermini = FALSE}.}

\item{proteinName}{character string with length = 1.
optional setting to include the name in the plot title.}

\item{printCitation}{Logical value. FALSE by default.
When \code{printCitation = TRUE} the citation for the pKa set is printed.
This allows for the user to easily obtain the dataset citation.
Will not print if there is a custom dataset.}

\item{...}{any additional parameters, especially those for plotting.}
}
\value{
If \code{plotResults = FALSE}, a dataframe
  is returned with the position, residue, and charge (-1 to +1). If
  \code{plotResults = TRUE}, a graphical output is returned (ggplot) showing
  the charge distribution.
}
\description{
This function will determine the charge of a peptide using the
  Henderson-Hasselbalch Equation. The output is a dataframe (default) or
  plot of charge calculations along the peptide sequence. Charges are
  determined globally, or along the entire chain.
}
\seealso{
\code{\link{pKaData}} for residue pKa values and
  \code{\link{hendersonHasselbalch}} for charge calculations.

Other charge functions: 
\code{\link{chargeCalculationLocal}()},
\code{\link{hendersonHasselbalch}()},
\code{\link{netCharge}()}
}
\concept{charge functions}
