% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargeCalculationLocal.R
\name{chargeCalculationLocal}
\alias{chargeCalculationLocal}
\title{Charge Calculation Along a Protein Sequence}
\usage{
chargeCalculationLocal(
  sequence,
  window = 7,
  pH = 7.2,
  pKaSet = "IPC_protein",
  printCitation = FALSE,
  plotResults = FALSE,
  proteinName = NA,
  ...
)
}
\arguments{
\item{sequence}{amino acid sequence as a single character string
or vector of single characters.
It also supports a single character string that specifies
the locaion of a .fasta or .fa file.}

\item{window}{a positive, odd integer. 7 by default.
Sets the size of sliding window, must be an odd number.
The window determines the number of residues to be analyzed and averaged
for each position along the sequence.}

\item{pH}{numeic value, 7.2 by default.
The envioronmental pH used to calculate residue charge.}

\item{pKaSet}{A character string or data frame. "IPC_protein" by default.
Character string to load specific, preloaded pKa sets.
 c("AA", "EMBOSS", "DTASelect", "Solomons", "Sillero", "Rodwell",
  "Lehninger", "Toseland", "Thurlkill", "Nozaki", "Dawson",
  "Bjellqvist", "ProMoST", "IPC_protein", "IPC_peptide")
 Alternativly, the user may supply a custom pKa dataset.
 The format must be a data frame where:
 Column 1 must be a character vector of residues AND
 Column 2 must be a numeric vector of pKa values.}

\item{printCitation}{Logical value. FALSE by default.
When \code{printCitation = TRUE} the citation for the pKa set is printed.
This allows for the user to easily obtain the dataset citation.
Will not print if there is a custom dataset.}

\item{plotResults}{logical value. TRUE by default.
If \code{plotResults = TRUE}, a ggplot of local charge is returned
If \code{plotResults = FALSE}, a dataframe of window charges are returned.}

\item{proteinName}{character string, optional. Used to add protein name
to the title in ggplot. Ignored if \code{plotResults = FALSE}.}

\item{...}{any additional parameters, especially those for plotting.}
}
\value{
see plotResults argument
}
\description{
This calculates the charge, as determined by the Henderson-Hasselbalch
  equation, for each window along the sequence. This function uses a
  sliding window. The output is either a graph or dataframe of calculations.
}
\seealso{
\code{\link{pKaData}} for residue pKa values and citations. See
  \code{\link{hendersonHasselbalch}} for charge calculations.

Other charge functions: 
\code{\link{chargeCalculationGlobal}()},
\code{\link{hendersonHasselbalch}()},
\code{\link{netCharge}()}
}
\concept{charge functions}
