% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCharge.R
\name{netCharge}
\alias{netCharge}
\title{Protein Charge Calculation, Net Charge}
\usage{
netCharge(
  sequence,
  pKaSet = "IPC_protein",
  pH = 7.2,
  includeTermini = TRUE,
  averaged = FALSE
)
}
\arguments{
\item{sequence}{amino acid sequence as a character string or vector of
individual residues. alternativly a character string of the path to a
.fasta / .fa file}

\item{pKaSet}{A character string or data frame. "IPC_protein" by default.
Character string to load specific, preloaded pKa sets.
 c("AA", "EMBOSS", "DTASelect", "Solomons", "Sillero", "Rodwell",
  "Lehninger", "Toseland", "Thurlkill", "Nozaki", "Dawson",
  "Bjellqvist", "ProMoST", "IPC_protein", "IPC_peptide")
 Alternativly, the user may supply a custom pKa dataset.
 The format must be a data frame where:
 Column 1 must be a character vector of residues AND
 Column 2 must be a numeric vector of pKa values.}

\item{pH}{numeic value, 7.2 by default.
The envioronmental pH used to calculate residue charge.}

\item{includeTermini}{Logical values, both TRUE by default. This
determines how the calculation handles the N- and C- terminus.
includeTermini determines if the calculation will use the charge of the
amine and carboxyl groups at the ends of the peptide (When TRUE). These
charges are ignored when \code{includeTermini = FALSE}. sumTermini
determines if the charge of the first (likely Met, therefore uncharged),
and final residue (varies) will be added to the termini charges, or if the
N and C terminus will be returned as seperate residues.
When \code{sumTermini = TRUE}, charges are summed. When
\code{sumTermini = FALSE}, the N and C terminus are added as a unique
residue in the DF. This will impact averages by increasing the sequence
length by 2. sumTermini is ignored if \code{includeTermini = FALSE}.}

\item{averaged}{logical value. FALSE by default.
When \code{averaged = FALSE}, the total net charge is returned.
When \code{averaged = TRUE}, the total net charge is averaged by the
sequence length. This gives a value of -1 to +1.}
}
\value{
numeric value. Either the net charge or average net charge, depending
  on the value of the averaged argument
}
\description{
This function will determine the net charge of a peptide using the
  Henderson-Hasselbalch Equation. The output is a numeric value. The output
  is the total net charge, but can be the average net charge.
}
\seealso{
\code{\link{pKaData}} for residue pKa values and citations. See
  \code{\link{hendersonHasselbalch}} for charge calculations.

Other charge functions: 
\code{\link{chargeCalculationGlobal}()},
\code{\link{chargeCalculationLocal}()},
\code{\link{hendersonHasselbalch}()}
}
\concept{charge functions}
