% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledHydropathyGlobal.R
\name{scaledHydropathyGlobal}
\alias{scaledHydropathyGlobal}
\title{Protein Scaled Hydropathy Calculations}
\usage{
scaledHydropathyGlobal(sequence, plotResults = FALSE, proteinName = NA, ...)
}
\arguments{
\item{sequence}{amino acid sequence as a single character string
or vector of single characters.
It also supports a single character string that specifies
the locaion of a .fasta or .fa file.}

\item{plotResults}{logical value, FALSE by default.
If \code{plotResults = TRUE} a plot will be the output.
If \code{plotResults = FALSE} the output is a data frame for each residue.}

\item{proteinName}{character string with length = 1.
optional setting to include the name in the plot title.}

\item{...}{any additional parameters, especially those for plotting.}
}
\value{
if \code{plotResults = TRUE}, a graphical representation data.
  Average is shown by the horizontal line.
  If \code{plotResults = FALSE}, a dataframe is reported
  with each amino acid and each residue value shown.
  Score for each residue shown in the column "Hydropathy".
}
\description{
This is used to calculate the scaled hydropathy of an amino acid
  sequence for each residue in the sequence.
  Results are based on the
}
\references{
Kyte, J., & Doolittle, R. F. (1982). A simple method for
  displaying the hydropathic character of a protein.
  Journal of molecular biology, 157(1), 105-132.
}
\seealso{
\code{\link{KDNorm}} for residue values.

Other scaled hydropathy functions: 
\code{\link{KDNorm}},
\code{\link{meanScaledHydropathy}()},
\code{\link{scaledHydropathyLocal}()}
}
\concept{scaled hydropathy functions}
