% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledHydropathyLocal.R
\name{scaledHydropathyLocal}
\alias{scaledHydropathyLocal}
\title{Calculate the Average Scaled Hydropathy of an Amino Acid Sequence}
\usage{
scaledHydropathyLocal(
  sequence,
  window = 9,
  plotResults = TRUE,
  proteinName = NA,
  ...
)
}
\arguments{
\item{sequence}{amino acid sequence as a single character string
or vector of single characters.
It also supports a single character string that specifies
the locaion of a .fasta or .fa file.}

\item{window}{a positive, odd integer. 7 by default.
Sets the size of sliding window, must be an odd number.
The window determines the number of residues to be analyzed and averaged
for each position along the sequence.}

\item{plotResults}{logical value, TRUE by default.
If \code{plotResults = TRUE} a plot will be the output.
If \code{plotResults = FALSE} the output is a data frame with scores for
each window analyzed.}

\item{proteinName}{character string with length = 1.
optional setting to replace the name of the plot if hydropathy = TRUE.}

\item{...}{any additional parameters, especially those for plotting.}
}
\value{
see plotResults argument
}
\description{
This is used to calculate the scaled hydropathy of an amino acid
  sequence using a sliding window. The output is either a dataframe or graph
  showing the calculated scores for each window along the sequence.
}
\references{
Kyte, J., & Doolittle, R. F. (1982). A simple method for
  displaying the hydropathic character of a protein.
  Journal of molecular biology, 157(1), 105-132.
}
\seealso{
\code{\link{KDNorm}} for residue values.

Other scaled hydropathy functions: 
\code{\link{KDNorm}},
\code{\link{meanScaledHydropathy}()},
\code{\link{scaledHydropathyGlobal}()}
}
\concept{scaled hydropathy functions}
