% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.check.R
\name{sequenceCheck}
\alias{sequenceCheck}
\title{Sequence Check Function}
\usage{
sequenceCheck(
  sequence,
  sequenceName = NA,
  method = "Stop",
  outputType = "String",
  outputName = NA,
  nonstandardResidues = NA,
  supressAAWarning = F,
  supressOutputMessage = F
)
}
\arguments{
\item{sequence}{amino acid sequence as a single character string
or vector of single characters.
It also supports a single character string that specifies
the locaion of a .fasta or .fa file.}

\item{sequenceName}{Optional setting.
\code{sequenceName = NA} for default naming.}

\item{method}{Required Setting.
 \code{method = c("Stop", "Warn", "Print")} "Stop" by default.
 "Stop" Reports invalid residues as an error and
prevents the function from continuing.
 "Warn" Reports invalid residues through a warning function.
 "Print" Reports invalid residues as a message,
 but doesn't halt the function.
 Any invalid sequences will be reported as intended.}

\item{outputType}{Required Setting. "String" By default.
\code{outputType = c("String", "Vector", "List", "None")}
Returns the sequence as a single string of amino acids.
"Vector" returns the sequence as a vector of individual characters
"List" returns the sequences as a list of individual characters
"None" prevents the function from returning a sequence.}

\item{outputName}{Optional Setting. "Sequence" by default.
Allows for a custom value name during assignment.}

\item{nonstandardResidues}{Optional setting.
Expands the amino acid alphabet.
NA or Character vector required.
Default values are "ACDEFGHIKLMNPQRSTVWY". Additional letters added here.
\code{nonstandardResidues = c("O,U")}
to allow Pyrrolysine (O) and Selenocysteine (U).}

\item{supressAAWarning}{If using nonstandardResidues,
 a warning will be issued.
set \code{nonstandardResidues = T}
to confirm addition of non-standard residues.}

\item{supressOutputMessage}{Set \code{supressOutputMessage = T}
to prevent sequence validity message}
}
\value{
A message and sequence are returned.
  If \code{supressOutputMessage = T}, the message is not returned.
  If \code{outputType = "None")}, the sequence is not returned.
  Otherwise, outputType will determine the format of the returned sequence.
  If the sequence contains an error, it will be reported
  based on the value of method.
  The Sequence will be assigned to the value "Sequence" if sequenceName
  is not specified. Otherwise the sequence is assigned to the value of
  sequenceName. This allows the sequences to be called by the user.
}
\description{
This is used validate a sequence of amino acids.
  It can additionally be used to load an amino acid sequence.
  It can also be used to coerce a sequence into a specific format.
}
