% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceMapCoordinates.R
\name{sequenceMapCoordinates}
\alias{sequenceMapCoordinates}
\title{Sequence Map Coordinates}
\usage{
sequenceMapCoordinates(sequence, nbResidues = 30)
}
\arguments{
\item{sequence}{amino acid sequence as a single character string
or vector of single characters.
It also supports a single character string that specifies
the locaion of a .fasta or .fa file.}

\item{nbResidues}{numeric value, 30 by default.
The number of residues to display on each row of the plot.
It is not reccomended to be over 50 or under 10 for standard sequences.
Optimal value may vary between sequences of extreme lengths.}
}
\value{
A data frame with rows containing the amino acid sequence, residue
  position within the sequence, as well as the row and column of each
  residue within the ggplot output of sequenceMap().
}
\description{
This is a function used to create a coordinate grid for the
  \code{\link{sequenceMap}} function. It is based on the length of the
  sequence being mapped, and how many residues per line are specified.
  The function wraps the sequence to have a number of columns that is
  the sequence length / number of residues per row, rounded up. \cr\cr
  This is intended for use within the sequenceMap function, however, this
  can also be used to identify the coordinates of residues within the ggplot
  coordinate plane for addition annotations.
}
\seealso{
\code{\link{sequenceMapCoordinates}} for mapping coordinates
}
