% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idpr.R
\docType{package}
\name{idpr}
\alias{idpr}
\title{idpr: profiling and analyzing Intrinsically Disordered Proteins in R}
\description{
idpr aims to integrate tools for the computational analysis of 
intrinsically disordered proteins (IDPs) within R. This package is used to 
identify known characteristics of IDPs for a sequence of interest with 
easily reported and dynamic results. Additionally, this package includes 
tools for IDP-based sequence analysis to be used in conjunction with other R 
packages. 
\cr
Please see the idpr vignettes for details on idpr functions and theory.
\code{browseVignettes("idpr")}
}
