% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupred.R
\name{iupred}
\alias{iupred}
\alias{iupredAnchor}
\alias{iupredRedox}
\title{Prediction of Intrinsic Disorder with IUPred2A}
\source{
Bálint Mészáros, Gábor Erdős, Zsuzsanna Dosztányi,
  IUPred2A: context-dependent prediction of protein disorder as a function of
  redox state and protein binding, Nucleic Acids Research, Volume 46, Issue
  W1, 2 July 2018, Pages W329–W337, \url{https://doi.org/10.1093/nar/gky384}
  \cr\cr
  Erdős, G., & Dosztányi, Z. (2020). Analyzing protein disorder with
  IUPred2A. Current Protocols in Bioinformatics, 70, e99.
  \url{https://doi.org/10.1002/cpbi.99}
}
\usage{
iupred(
  uniprotAccession,
  iupredType = "long",
  plotResults = TRUE,
  proteinName = NA
)

iupredAnchor(uniprotAccession, plotResults = TRUE, proteinName = NA)

iupredRedox(uniprotAccession, plotResults = TRUE, proteinName = NA)
}
\arguments{
\item{uniprotAccession}{character string specifying the UniProt Accession
of the protein of interest. Used to fetch predictions from IUPreds REST API}

\item{iupredType}{character string. "long" by default. accepted types are
c("long", "short", "glob"). See "Prediction Type" information below.}

\item{plotResults}{logical value. TRUE by default.
If \code{plotResults = TRUE}, a ggplot of IUPred predictions is returned
If \code{plotResults = FALSE}, a data frame of predictions is returned.}

\item{proteinName}{character string, optional. Used to add protein name
to the title in ggplot. Ignored if \code{plotResults = FALSE}.}
}
\value{
see plotResults argument.
}
\description{
This function makes a connection to the IUPred2A REST API based on the type
  of analysis and UniProt accession number. This requires the user to know
  the accession number of their protein and a connection to the internet.
  The results are then formatted to match output in the idpr package. \cr \cr
  Predictions are made on a scale of 0-1, where any residues with a score
  over 0.5 are predicted to be disordered, and any residue scoring below 0.5
  are predicted to be ordered (when using "long" and "short" predictions).\cr
  The output is either a graph (ggplot) or data frame of predictions.
  \cr\cr
  \strong{iupred()} is used for standard predictions of intrinsic disorder
  of an amino acid sequence. This is the core of predictions.
  Predictions vary by iupredType (details below)
  The results are either a ggplot or data frame of the fetched IUPred2.
  predictions.
  \cr
  \strong{iupredAnchor()} is used to combine the output of IUPred2 long with
  ANCHOR2 predictions. ANCHOR2 is a context-dependent predictor of binding
  regions for protein-protein interactions. The results are either a ggplot
  with 2 lines, one for IUPred2 long and another for ANCHOR predictions, or
  a data frame with both IUPred2 long and ANCHOR Predictions. Values are
  fetched by the IUPred2A REST API.
  \cr
  \strong{iupredRedox()} is used to predict redox-sensitive regions that may
  experience induced folding upon changing environments.
  This is a context-dependent predictor of disordered regions depending on
  a reducing (plus) or oxidizing (minus) environment. The results can be
  a ggplot with two IUPred2 long predictions, one for plus and another for
  minus environments, with redox sensitive regions shaded (if predicted).
  Alternatively, the results can be a data frame with both IUPred2 long plus
  and minus predictions as well as a column of logical values where a residue
  that is TRUE is predicted to be in a redox sensitive region. Values are
  fetched by the IUPred2A REST API.
  \cr \cr
  IUPred2 website is located at \url{https://iupred2a.elte.hu/}.
  For detailed information on using IUPred2A, please refer to
  \href{https://doi.org/10.1002/cpbi.99}{Erdős & Dosztány (2020)}
  Analyzing protein disorder with IUPred2A.
  Current Protocols in Bioinformatics, 70, e99.
  Additionally, please see
  \href{https://doi.org/10.1093/nar/gky384}{Mészáros et al (2019)}
  for further information, theory, and applications of IUPred2A.
  \cr \cr
  \strong{Please cite these articles if you use any iupred function.}
}
\section{Prediction Type}{

  Information from \url{https://iupred2a.elte.hu/help_new} on 5.22.20
  Additionally, see the sources for further details and source information.
  This is only relevant for iupred(). iupredAnchor() and iupredRedox()
  always utilize "long" for data in the REST API.
  \itemize{
    \item Long predictions of disorder (Default)
       \itemize{
         \item when iupredType = "long"
         \item Optimized for global predictions of disorder, specifically
           disordered regions over 30 amino acids in length.
         \item "long" is always used for iupredAnchor() and iupredRedox().
       }
     \item Short predictions of disorder
       \itemize{
         \item when iupredType = "short"
         \item Best for predicting small regions of disorder, especially
           in mostly structured proteins.
         \item Has adjustments for termini, since sequence ends are often
           disordered.
       }
     \item Structured predictions
       \itemize{
         \item when iupredType = "glob"
         \item Used to predict regions of globular folding.
         \item please see
           \href{https://doi.org/10.1002/cpbi.99}{Erdős & Dosztány (2020)}
           for further information on interpreting these results.
       }
   }
}

\examples{
#A UniProt Accession must be specified.
##this example uses human P53.
TP53_UniProt <- "P04637"
\dontrun{
#Getting data as a data frame
exampleDF_long <- iupred(uniprotAccession = TP53_UniProt,
                         iupredType = "long",
                         plotResults = FALSE)
head(exampleDF_long)

exampleDF_short <- iupred(uniprotAccession = TP53_UniProt,
                          iupredType = "short",
                          plotResults = FALSE)
head(exampleDF_short)

exampleDF_anchor <- iupredAnchor(uniprotAccession = TP53_UniProt,
                                 plotResults = FALSE)
head(exampleDF_anchor)

exampleDF_redox <- iupredRedox(uniprotAccession = TP53_UniProt,
                              plotResults = FALSE)
head(exampleDF_redox)

#Plotting

iupred(uniprotAccession = TP53_UniProt,
    iupredType = "long",
    plotResults = TRUE)

iupred(uniprotAccession = TP53_UniProt,
    iupredType = "short",
    plotResults = TRUE)

iupredAnchor(uniprotAccession = TP53_UniProt,
          plotResults = TRUE)

iupredRedox(uniprotAccession = TP53_UniProt,
         plotResults = TRUE)
}
#A valid internet connection is needed to make
##A connection with the IUPred REST API
}
