% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuralTendencyPlot.R
\name{structuralTendencyPlot}
\alias{structuralTendencyPlot}
\title{Plotting Structural Tendency of Amino Acid Sequence}
\usage{
structuralTendencyPlot(
  sequence,
  graphType = "pie",
  summarize = FALSE,
  proteinName = NA,
  alphabetical = FALSE,
  disorderPromoting = c("P", "E", "S", "Q", "K", "A", "G"),
  disorderNeutral = c("D", "T", "R"),
  orderPromoting = c("M", "N", "V", "H", "L", "F", "Y", "I", "W", "C"),
  ...
)
}
\arguments{
\item{sequence}{amino acid sequence (or pathway to a fasta file)
as a character string. Supports multiple sequences / files, as a
character vector of strings.}

\item{graphType}{character string, required.
graphType must be set to c("pie", "bar", "none").
When \code{graphType = "pie"}, the output is a pie chart.
When \code{graphType = "bar"}, the output is a bar chart.
When \code{graphType = "none"}, the output is the dataframe that would
otherwise be used to plot the data.}

\item{summarize}{logical value, FALSE by default.
When \code{summarize = TRUE}, each residue is agregated into Disorder
Tendancy Groups. (See \code{\link{structuralTendency}} for more details).
When \code{summarize = FALSE}, residue identity is preserved and
the output is colored by Disorder Tendancy Groups.}

\item{proteinName, }{optional character string. NA by default.
Used to either add the name of the protein to the plot title.}

\item{alphabetical}{logical value, FALSE by default.
Order of residues on plot axis. Only relevant when
 \code{summarize = FALSE}, otherwise is ignored.
 If FALSE, ordering is grouped by Disorder Tendancy (P, E, S, ..., W, C).
 If TRUE, the residues are ordered alphabetically (A, C, D, E, ..., W, Y).}

\item{disorderPromoting, disorderNeutral, orderPromoting}{character vectors
of individual residues to be matched with the input sequence. Defaults:
\itemize{
  \item disorderPromoting = c("P", "E", "S", "Q", "K", "A", "G")
  \item orderPromoting =
     c("M", "N", "V", "H", "L", "F", "Y", "I", "W", "C")
  \item disorderNeutral = c("D", "T", "R")
}
It is not reccomended to change these.}

\item{...}{additional arguments to be passed to
\code{\link{structuralTendency}} and
\code{\link[ggplot2]{ggplot}}}
}
\value{
a dataframe containing each residue from the sequence
  matched with its structural tendancy, defined by disorderPromoting,
  disorderNeutral, and orderPromoting.
}
\description{
Convenient graphing for the \code{\link{structuralTendency}} function.
}
\references{
Kulkarni, Prakash, and Vladimir N. Uversky. "Intrinsically
  disordered proteins: the dark horse of the dark proteome."
  Proteomics 18.21-22 (2018): 1800061.
  \url{https://doi.org/10.1002/pmic.201800061}.
}
\seealso{
Other structural tendency: 
\code{\link{structuralTendency}()}
}
\concept{structural tendency}
